+++ src/Navigation.tsx
@@ -6,9 +6,10 @@ const Navigation = () => {
       <nav className="navbar navbar-expand-lg bg-light">
         <div className="container-fluid">
           <div className="collapse navbar-collapse" id="navbarSupportedContent">
-            <Link to="/">Azure-sample</Link> |
-            <Link to="/repo?name=js-e2e"> Repos</Link> |
-            <Link to="/summary"> Summary</Link>
+            <Link to="/">Azure-sample</Link> &nbsp;|&nbsp;
+            <Link to="/orgmeta">Good & Bad</Link> &nbsp;|&nbsp;
+            <Link to="/repo?name=js-e2e">Repos</Link> &nbsp;|&nbsp;
+            <Link to="/summary">Summary</Link>&nbsp;
           </div>
         </div>
       </nav>