+++ src/utilities/types.ts
@@ -1,0 +1,4 @@
+export type IWeightedRepo = {
+    repositoryName: string
+    weight: number
+}